* WAL refactoring

+++ src/kv/iwal.c
@@ -21,40 +21,40 @@ typedef enum {
 
 #pragma pack(push, 1)
 typedef struct WBSEP {
-  uint8_t id;
+  uint32_t id;
   uint32_t crc;
   uint32_t len;
 } WBSEP;
 
 typedef struct WBSET {
-  uint8_t id;
-  uint8_t val;
+  uint32_t id;
+  uint32_t val;
   off_t off;
   off_t len;
 } WBSET;
 
 typedef struct WBCOPY {
-  uint8_t id;
+  uint32_t id;
   off_t off;
   off_t len;
   off_t noff;
 } WBCOPY;
 
 typedef struct WBWRITE {
-  uint8_t id;
+  uint32_t id;
   uint32_t crc;
   uint32_t len;
   off_t off;
 } WBWRITE;
 
 typedef struct WBRESIZE {
-  uint8_t id;
+  uint32_t id;
   off_t osize;
   off_t nsize;
 } WBRESIZE;
 
 typedef struct WBFIXPOINT {
-  uint8_t id;
+  uint32_t id;
   uint64_t ts;
 } WBFIXPOINT;
 #pragma pack(pop)
